Include the Field Id
When we need to update/insert new table record, is necessary to know the field id. Most of the cases the id is the name of the field in lower cases, but sometimes, the filed id is named like "o_qu" which is the first field when is created.
Without know, exactly the field id is not possible to create new records nor updates.